+++ gwas/gwas.py
@@ -79,6 +79,7 @@ def parse_args(args):
     pheno_parser.add_argument("--dict-sep", default=None, type=str, choices=[None, ',', 'comma', ';', 'semicolon', '\t', 'tab', ' ', 'space', 'delim-whitespace'],
         help="Delimiter to use when reading --dict-file; by default uses delimiter provided to --pheno-sep")
     pheno_parser.add_argument("--fam", type=str, default=None, help="an argument pointing to a plink's .fam file, use by gwas.py script to pre-filter phenotype information (--pheno) with the set of individuals available in the genetic file (--geno-file / --geno-fit-file). Optional when either --geno-file (or --geno-fit-file) is in plink's format, otherwise required - but IID in this file must be consistent with identifiers of the genetic file.")
+    pheno_parser.add_argument("--bim", type=str, default=None, help="an optional argument pointing to a plink's .bim file, used by gwas.py script whenever it needs to know genomic positions or marker names (for example, this is needed when SAIGE GWAS uses --chunk-size-bp option)")
     pheno_parser.add_argument("--pheno", type=str, default=[], nargs='+', help="target phenotypes to run GWAS (must be columns of the --pheno-file")
     pheno_parser.add_argument("--covar", type=str, default=[], nargs='+', help="covariates to control for (must be columns of the --pheno-file); individuals with missing values for any covariates will be excluded not just from <out>.covar, but also from <out>.pheno file")
     pheno_parser.add_argument("--variance-standardize", type=str, default=None, nargs='*', help="the list of continuous phenotypes to standardize variance; accept the list of columns from the --pheno file (if empty, applied to all); doesn't apply to dummy variables derived from NOMINAL or BINARY covariates.")
@@ -120,6 +121,7 @@ def parser_gwas_add_arguments(args, func, parser):
     parser.add_argument("--geno-fit-file", type=str, default=None, help="genetic file to use in a first stage of mixed effect model. Expected to have the same set of individuals as --geno-file (this is NOT validated by the gwas.py script, and it is your responsibility to follow this assumption). Optional for standard association analysis (e.g. if for plink's glm). The argument supports the same file types as the --geno-file argument. Noes not support '@' (because mixed effect tools typically expect a single file at the first stage.")
     parser.add_argument("--chr2use", type=str, default='1-22', help="Chromosome ids to use "
          "(e.g. 1,2,3 or 1-4,12,16-20). Used when '@' is present in --geno-file, and allows to specify for which chromosomes to run the association testing.")
+    parser.add_argument("--chunk-size-bp", type=int, default=None, help="chunk size (in base pairs); used when GWAS is done with SAIGE; require --bim argument")
 
     parser.add_argument('--analysis', type=str, default=['plink2', 'figures'],
         nargs='+', choices=['plink2', 'regenie', 'saige', 'figures'],
@@ -146,6 +148,7 @@ def parser_merge_saige_add_arguments(args, func, parser):
     parser.add_argument("--sumstats", type=str, default=None, help="sumstat file produced by plink2, containing @ as chromosome label place holder")
     parser.add_argument("--basename", type=str, default=None, help="basename for .vmiss, .afreq and .hardy files, with @ as chromosome label place holder")
     parser.add_argument("--chr2use", type=str, default='1-22', help="Chromosome ids to use, (e.g. 1,2,3 or 1-4,12,16-20).")
+    parser.add_argument("--chunks", type=int, default=None, help="How many chunks to merge; mutually exclusive with --chr2use")
     parser.set_defaults(func=func)
 
 def extract_variables(df, variables, pheno_dict_map, log):
@@ -237,6 +240,9 @@ def fix_and_validate_gwas_args(args, log):
     if ('saige' in args.analysis) and (not args.geno_fit_file):
         raise ValueError('--geno-fit-file must be specified for --analysis saige')
 
+    if ('saige' in args.analysis) and (args.chunk_size_bp is not None) and (not args.bim):
+        raise ValueError('--bim must be specified together with --chunk-size-bp in case of --analysis saige')
+
 def make_regenie_commands(args, logistic, step):
     geno_fit_file = args.geno_fit_file
     geno_file = args.geno_file.replace('@', '${SLURM_ARRAY_TASK_ID}')
@@ -273,12 +279,42 @@ def make_saige_commands(args, logistic, step):
     geno_fit_file = args.geno_fit_file
     geno_file = args.geno_file.replace('@', '${SLURM_ARRAY_TASK_ID}')
 
+    use_chunks = (args.chunk_size_bp is not None)
+    if use_chunks:
+        bim = read_bim(args, args.bim)
+        chunk_def = {'chr':[], 'chunk':[], 'num_snps_in_chunk':[], 'a':[], 'b':[]}
+        chunk_index = 0
+        for chr_label in args.chr2use:
+            bim_chr = bim[bim['CHR'].astype(str) == chr_label]
+
+            min_bp = 1e6 * int(bim_chr.BP.min()/1e6)
+            max_bp = bim_chr.BP.max() + 1
+
+            chunks=[(x, x+args.chunk_size_bp) for x in range(int(min_bp), int(max_bp), int(args.chunk_size_bp))]
+            for (a, b) in chunks:
+                num_snps = np.sum((bim_chr.BP > a) & (bim_chr.BP <= b))
+                if num_snps == 0: continue
+                chunk_index += 1
+                chunk_def['chr'].append(chr_label)
+                chunk_def['chunk'].append(chunk_index)
+                chunk_def['a'].append(a)
+                chunk_def['b'].append(b)
+                chunk_def['num_snps_in_chunk'].append(num_snps)
+
+        chrom_def = ' '.join(['[{}]={}'.format(index+1, value) for index, value in enumerate(chunk_def['chr'])])
+        start_def = ' '.join(['[{}]={}'.format(index+1, value+1) for index, value in enumerate(chunk_def['a'])])
+        end_def = ' '.join(['[{}]={}'.format(index+1, value) for index, value in enumerate(chunk_def['b'])])
+        snps_def = ' '.join(['[{}]={}'.format(index+1, value) for index, value in enumerate(chunk_def['num_snps_in_chunk'])])
+        array_spec = [str(x) for x in range(1, len(chunk_def['chr']) + 1)]
+    else:
+        array_spec = args.chr2use
+
     if ('@' in geno_fit_file): raise(ValueError('--geno-fit-file contains "@", hense it is incompatible with SAIGE step1 which require a single file'))
     if (is_pgen_file(geno_fit_file) or is_pgen_file(geno_file)): raise(ValueError('--geno-file / --geno-fit-file can not point to plink2 pgen file for SAIGE analysis'))
 
     cmd_step1 = ''; cmd_step2 = ''
 
-    for pheno in args.pheno:
+    for pheno_index, pheno in enumerate(args.pheno):
         cmd_step1 += '\n$SAIGE step1_fitNULLGLMM.R ' + \
             (" --plinkFile={} ".format(remove_suffix(geno_fit_file, '.bed')) if is_bed_file(geno_fit_file) else "") + \
             (" --vcfFile={f} --vcfFileIndex={f}.tbi --vcfField={vf} ".format(f=geno_fit_file, vf=args.vcf_field) if is_vcf_file(geno_fit_file) else "") + \
@@ -294,23 +330,34 @@ def make_saige_commands(args, logistic, step):
             " --minMAFforGRM=0.01" + \
             " --tauInit=1,0 "
 
+        if use_chunks and (pheno_index==0):
+            cmd_step2 += f"\ndeclare -A CHUNKS_CHR=({chrom_def})"
+            cmd_step2 += f"\ndeclare -A CHUNKS_START=({start_def})"
+            cmd_step2 += f"\ndeclare -A CHUNKS_END=({end_def})"
+            cmd_step2 += f"\ndeclare -A CHUNKS_SNPS=({snps_def})"
+            cmd_step2 += "\n"
+
         cmd_step2 += '\n$SAIGE step2_SPAtests.R ' + \
             (" --plinkFile={} ".format(remove_suffix(geno_file, '.bed')) if is_bed_file(geno_file) else "") + \
             (" --vcfFile={f} --vcfFileIndex={f}.tbi --vcfField={vf} ".format(f=geno_file, vf=args.vcf_field) if is_vcf_file(geno_file) else "") + \
             (" --bgenFile={f} --bgenFileIndex={f}.bgi ".format(f=geno_file) if is_bgen_file(geno_file) else "") + \
-            " --chr ${SLURM_ARRAY_TASK_ID}" + \
+            (" --chr ${SLURM_ARRAY_TASK_ID}" if (not use_chunks) else "") + \
+            (" --chr ${CHUNKS_CHR[${SLURM_ARRAY_TASK_ID}]}" if (use_chunks) else "") + \
+            (" --start ${CHUNKS_START[${SLURM_ARRAY_TASK_ID}]}" if (use_chunks) else "") + \
+            (" --end ${CHUNKS_END[${SLURM_ARRAY_TASK_ID}]}" if (use_chunks) else "") + \
             (" --minMAF={}".format(args.maf) if (args.maf is not None) else '') + \
             " --minMAC=1" + \
             " --sampleFile={}.sample".format(args.out) + \
             " --GMMATmodelFile={}_{}.step1.rda".format(args.out, pheno) + \
             " --varianceRatioFile={}_{}.step1.varianceRatio.txt".format(args.out, pheno) + \
-            " --SAIGEOutputFile={}_chr${{SLURM_ARRAY_TASK_ID}}_{}.saige".format(args.out, pheno) + \
+            (" --SAIGEOutputFile={}_chunk${{SLURM_ARRAY_TASK_ID}}_{}.saige".format(args.out, pheno) if use_chunks else "") + \
+            (" --SAIGEOutputFile={}_chr${{SLURM_ARRAY_TASK_ID}}_{}.saige".format(args.out, pheno) if not use_chunks else "") + \
             " --numLinesOutput=2" + \
             (" --IsOutputAFinCaseCtrl=TRUE" if logistic else '')+ \
             (" --IsOutputNinCaseCtrl=TRUE" if logistic else '') + \
             " --LOCO=TRUE "
 
-    return cmd_step1 if step==1 else cmd_step2
+    return (cmd_step1, None) if step==1 else (cmd_step2, array_spec)
 
 # this function works similarly to ** in python:
 # all args from args_list that are not None are passed to the caller
@@ -347,15 +394,18 @@ def make_regenie_merge_commands(args, logistic):
             '\n'
     return cmd
 
-def make_saige_merge_commands(args, logistic):
+def make_saige_merge_commands(args, logistic, array_spec):
     cmd = ''
+    use_chunks = (args.chunk_size_bp is not None)
     for pheno in args.pheno:
         cmd += '$PYTHON gwas.py merge-saige ' + \
             pass_arguments_along(args, ['info-file', 'info', 'maf', 'hwe', 'geno']) + \
-            ' --sumstats {out}_chr@_{pheno}.saige'.format(out=args.out, pheno=pheno) + \
+            (' --sumstats {out}_chr@_{pheno}.saige'.format(out=args.out, pheno=pheno) if (not use_chunks) else "") + \
+            (' --sumstats {out}_chunk@_{pheno}.saige'.format(out=args.out, pheno=pheno) if use_chunks else "") + \
             ' --basename {out}_chr@'.format(out=args.out) + \
             ' --out {out}_{pheno} '.format(out=args.out, pheno=pheno) + \
             ' --chr2use {} '.format(','.join(args.chr2use)) + \
+            (' --chunks {}'.format(len(array_spec)) if use_chunks else "") + \
             '\n'
     return cmd
 
@@ -416,7 +466,7 @@ def make_prsice2_commands(args, logistic):
 
     return cmd
 
-def make_slurm_header(args, array=False):
+def make_slurm_header(args, array_spec=None):
     return """#!/bin/bash
 #SBATCH --job-name={job_name}
 #SBATCH --account={account}
@@ -438,7 +488,7 @@ export RSCRIPT="singularity exec --home $PWD:/home $SIF/r.sif Rscript"
 export PRSICE2="singularity exec --home $PWD:/home $SIF/gwas.sif PRSice_linux"
 export SAIGE="singularity exec --home $PWD:/home $SIF/saige.sif"
 
-""".format(array="#SBATCH --array={}".format(','.join(args.chr2use)) if array else "",
+""".format(array="#SBATCH --array={}".format(','.join(array_spec)) if (array_spec is not None) else "",
            modules = '\n'.join(['module load {}'.format(x) for x in args.config['slurm']['module_load']]),
            job_name = args.config['slurm']['job_name'],
            account = args.config['slurm']['account'],
@@ -561,7 +611,7 @@ def execute_pgrs(args, log):
 
     if 'prsice2' in args.analysis:
         commands = [make_prsice2_commands(args, logistic)]
-        num_jobs = append_job(args, commands, True, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, args.chr2use, num_jobs+1, cmd_file, submit_jobs)
 
     log.log('To submit all jobs via SLURM, use the following scripts, otherwise execute commands from {}'.format(cmd_file))
     print('\n'.join(submit_jobs))
@@ -586,35 +636,41 @@ def execute_gwas(args, log):
     if 'plink2' in args.analysis:
         commands = [make_plink2_info_commands(args),
                     make_plink2_glm_commands(args, logistic)]
-        num_jobs = append_job(args, commands, True, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, args.chr2use, num_jobs+1, cmd_file, submit_jobs)
 
         commands = [make_plink2_merge_commands(args, logistic)]
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, None, num_jobs+1, cmd_file, submit_jobs)
 
     if 'regenie' in args.analysis:
         commands = [make_regenie_commands(args, logistic, step=1)]
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, None, num_jobs+1, cmd_file, submit_jobs)
 
         commands = [make_plink2_info_commands(args), make_regenie_commands(args, logistic, step=2)]
-        num_jobs = append_job(args, commands, True, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, args.chr2use, num_jobs+1, cmd_file, submit_jobs)
 
         commands = [make_regenie_merge_commands(args, logistic)]
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, None, num_jobs+1, cmd_file, submit_jobs)
 
     if 'saige' in args.analysis:
-        commands = [make_saige_commands(args, logistic, step=1)]
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        command, _ = make_saige_commands(args, logistic, step=1)
+        num_jobs = append_job(args, [command], None, num_jobs+1, cmd_file, submit_jobs)
 
-        commands = [make_plink2_info_commands(args), make_saige_commands(args, logistic, step=2)]
-        num_jobs = append_job(args, commands, True, num_jobs+1, cmd_file, submit_jobs)
+        cpus_per_task = args.config['slurm']['cpus_per_task']
+        args.config['slurm']['cpus_per_task'] = args.config['saige']['cpus_per_task_stage2']
+        command, array_spec = make_saige_commands(args, logistic, step=2)
+        num_jobs = append_job(args, [command], array_spec, num_jobs+1, cmd_file, submit_jobs)
+        args.config['slurm']['cpus_per_task'] = cpus_per_task
 
-        commands = [make_saige_merge_commands(args, logistic)]
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        commands = [make_plink2_info_commands(args)]
+        num_jobs = append_job(args, commands, args.chr2use, num_jobs+1, cmd_file, submit_jobs)
+
+        commands = [make_saige_merge_commands(args, logistic, array_spec)]
+        num_jobs = append_job(args, commands, None, num_jobs+1, cmd_file, submit_jobs)
 
     commands = []
     if 'figures' in args.analysis: commands.append(make_figures_commands(args))
     if commands:
-        num_jobs = append_job(args, commands, False, num_jobs+1, cmd_file, submit_jobs)
+        num_jobs = append_job(args, commands, None, num_jobs+1, cmd_file, submit_jobs)
 
     log.log('To submit all jobs via SLURM, use the following scripts, otherwise execute commands from {}'.format(cmd_file))
     print('\n'.join(submit_jobs))
@@ -735,8 +791,9 @@ def merge_saige(args, log):
 
     fix_and_validate_chr2use(args, log)
     pattern = args.sumstats
-    check_input_file(pattern, args.chr2use)
-    df=pd.concat([pd.read_csv(pattern.replace('@', chri), delim_whitespace=True) for chri in args.chr2use])
+    chr_or_chunk_labels = list(range(1, args.chunks + 1)) if args.chunks else args.chr2use
+    check_input_file(pattern, chr_or_chunk_labels)
+    df=pd.concat([pd.read_csv(pattern.replace('@', str(label)), delim_whitespace=True) for label in chr_or_chunk_labels])
     cols = [('SNPID', 'SNP'), ('CHR', 'CHR'), ('POS', 'BP'),
             ('Allele1', 'A2'), ('Allele2', 'A1'),
             ('N', 'N'), ('BETA', 'BETA'), ('SE', 'SE'), 
@@ -811,6 +868,12 @@ class Logger(object):
             with open(self.fh + '.error', 'w') as error_fh:
                 error_fh.write(str(msg).rstrip() + '\n')
 
+def read_bim(args, bim_file):
+    log.log('reading {}...'.format(bim_file))
+    bim = pd.read_csv(bim_file, delim_whitespace=True, header=None, names='CHR SNP GP BP A1 A2'.split())
+    log.log('done, {} rows, {} cols'.format(len(bim), bim.shape[1]))
+    return bim
+
 def read_fam(args, fam_file):
     log.log('reading {}...'.format(fam_file))
     fam = pd.read_csv(fam_file, delim_whitespace=True, header=None, names='FID IID FatherID MotherID SEX PHENO'.split(), dtype=str)
@@ -822,20 +885,20 @@ def read_fam(args, fam_file):
 # Intput:
 #   args - command line arguments (as usual)
 #   commands - array of commands to put in a SLURM job
-#   as_array - whether the job is an array of 22 jobs, or not
+#   array_spec - list of SLURM_ARRAY_TASK_ID if job array, otherwise None
 # Output:
 #   slurm_job_file - filename of a .job file to output commands (with SLURM header)
 #   cmd_file - file that concatenate all commands (in case user wants to execute everything by a BASH on a local node)
 #   submit_job_list - instructions for a user about how to schedule SLURM jobs
-def append_job(args, commands, as_array, slurm_job_index, cmd_file, submit_jobs_list):
+def append_job(args, commands, array_spec, slurm_job_index, cmd_file, submit_jobs_list):
     slurm_job_file = args.out + '.{}.job'.format(slurm_job_index)
     with open(slurm_job_file, 'w') as f:
-        f.write(make_slurm_header(args, array=as_array) + '\n'.join(commands) + '\n')
+        f.write(make_slurm_header(args, array_spec=array_spec) + '\n'.join(commands) + '\n')
     
     with open(cmd_file, 'a') as f:
         for command in commands:
-            if as_array:
-                f.write('for SLURM_ARRAY_TASK_ID in {}; do {}; done\n'.format(' '.join(args.chr2use), command))
+            if array_spec is not None:
+                f.write('for SLURM_ARRAY_TASK_ID in {}; do {}; done\n'.format(' '.join(array_spec), command))
             else:
                 f.write('{}\n'.format(command))